separating tablesample from utilities
