+++ verticapy/core/__init__.py
@@ -14,4 +14,3 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.core.vdataframe import vDataFrame
